+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Created By: BENGUERGOURA Oussama & HADJERSI Mohamed | 20/10/2024
 
 ### Added
 
-* Creating `BarSide`component with multiple `BarSideMenu` and `BarSideItem`. [GTS-PRINT#32](https://quire.io/w/GTS-PRINT31/2)
+* Adding `BarSide` test suites. [GTS-PRINT#6](https://quire.io/w/GTS-PRINT31/6)
+* Creating `BarSide`component with multiple `BarSideMenu` and `BarSideItem`. [GTS-PRINT#2](https://quire.io/w/GTS-PRINT31/2)
 
 ### Changed